Add files via upload
The method used for optimizing the value of C in this code is called Bayesian optimization. More specifically, this code is using Gaussian process-based Bayesian optimization, which is a variant of Bayesian optimization that uses a Gaussian process to model the objective function and generate the next set of points to evaluate. The gp_minimize function from the scikit-optimize library is used to perform the optimization.

In Bayesian optimization, the goal is to find the input (in this case, the value of C) that maximizes the output of an unknown, black-box function (in this case, the test accuracy of the model). The method works by iteratively selecting points in the input space to evaluate, based on a probabilistic model of the objective function that is updated as more points are evaluated. The model is used to balance exploration (trying inputs in new regions of the input space) and exploitation (trying inputs that are likely to be optimal based on the model) in order to find the global optimum of the function with as few evaluations as possible.